Move fuzzy completion logic into a separate module
